Refactor gSwap API into modular helpers
